added devise, pundit policies and updated the code

+++ app/controllers/articles_controller.rb
@@ -1,8 +1,6 @@
 class ArticlesController < ApplicationController
   before_action :find_article, only: [:show, :edit, :update, :destroy]
-  before_action :require_user, except: [:show, :index]
-  before_action :require_same_user, only: [:edit, :update, :destroy]
-
+  
   def show
   end
 
@@ -15,6 +13,7 @@ class ArticlesController < ApplicationController
   end
 
   def edit
+    authorize @article
   end
 
   def create
@@ -29,6 +28,7 @@ class ArticlesController < ApplicationController
   end
 
   def update
+    authorize @article
     if @article.update(article_params)
       flash[:notice] = "Article was updated successfully!"
       redirect_to @article
@@ -38,6 +38,7 @@ class ArticlesController < ApplicationController
   end
 
   def destroy
+    authorize @article
     if @article.destroy
       flash[:notice] = "Article destroyed successfully!"
       redirect_to articles_path
@@ -54,13 +55,6 @@ class ArticlesController < ApplicationController
 
   def article_params
     params.require(:article).permit(:title, :description, category_ids: [])
-  end
-
-  def require_same_user
-    if current_user != @article.user && !current_user.admin?
-      flash[:alert] = "You can only edit or delete only your own article" 
-      redirect_to @article
-    end
   end
 
 end
